rename modal_analysis to pastis_analysis
